FlightManager and SeatingQueue updated. Minor changes to AirlineInterface and FlightData
FlightManager and SeatingQueue should be working now, just need the
flightmap to calculate mileagles.

Fixed a few minor issues in AirlineInterface that were giving me
compiler errors.

Commented out removePassenger from FlightData. The erase function it
calls does not appear to be a member of PassengerData, and was giving a
compiler error.